# PERF-#6378: use numpy.array operations in internals of iloc/loc operation