@anmyachev 2023-07-17
New results:

<details>

```bash
       before           after         ratio                                                                                                                                                                                         [61/1743]
     [ea0a6acb]       [48f117bf]
     <master>
+      3.570.2ms       4.110.1ms     1.15  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       3.420.1ms       3.850.1ms    ~1.13  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'python_list_take_10_values')
       2.850.1ms      3.210.06ms    ~1.13  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       21.40.9ms         24.01ms    ~1.12  benchmarks.TimeIndexingColumns.time_iloc([1000000, 10])
       16.90.9ms         18.51ms     1.09  benchmarks.TimeIndexingColumns.time___getitem__([5000, 5000])
       3.740.2ms       4.080.1ms     1.09  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       3.690.2ms       4.020.2ms     1.09  benchmarks.TimeIndexing.time_loc([1000000, 10], 'python_list_take_10_values')
       3.960.2ms       4.240.1ms     1.07  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
         55020ms         59010ms     1.07  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool_series')
       5.050.1ms       5.390.2ms     1.07  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
      2.940.07ms      3.130.08ms     1.07  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       5.270.2ms       5.620.3ms     1.07  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       65.20.5ms         68.62ms     1.05  benchmarks.TimeIndexing.time_loc([1000000, 10], 'numpy_array_take_all_values')
       3.330.1ms       3.500.1ms     1.05  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
         55310ms         57720ms     1.05  benchmarks.TimeIndexing.time_loc([5000, 5000], 'slice')
       55.50.6ms         57.91ms     1.04  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_array')
       20.20.5ms       21.00.5ms     1.04  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'scalar')
      3.600.06ms      3.750.06ms     1.04  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       3.770.1ms       3.930.1ms     1.04  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       16.90.2ms       17.60.4ms     1.04  benchmarks.TimeIndexingColumns.time_iloc([5000, 5000])
       3.410.1ms       3.540.2ms     1.04  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
      3.260.06ms       3.370.2ms     1.03  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       8.050.1ms      8.300.09ms     1.03  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
         23.71ms         24.22ms     1.02  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'bool_series')
       24.00.4ms       24.60.2ms     1.02  benchmarks.TimeIndexing.time_loc([5000, 5000], 'numpy_array_take_all_values')
         37.82ms         38.61ms     1.02  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_series')
       22.30.8ms       22.80.7ms     1.02  benchmarks.TimeIndexingColumns.time___getitem__([1000000, 10])
         52820ms         53920ms     1.02  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool_series')
          3337ms          3396ms     1.02  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       17.90.4ms       18.20.7ms     1.02  benchmarks.TimeIndexing.time_loc([5000, 5000], 'python_list_take_10_values')
      5.540.06ms       5.640.1ms     1.02  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       4.240.1ms      4.310.04ms     1.02  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'scalar')
         48710ms          4955ms     1.02  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
      3.400.07ms       3.460.2ms     1.02  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
         58010ms         58920ms     1.02  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'bool_array')
      3.020.03ms      3.070.06ms     1.01  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       4.150.3ms       4.210.1ms     1.01  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
          4926ms          4996ms     1.01  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
      3.890.07ms       3.950.2ms     1.01  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
         47.01ms         47.61ms     1.01  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'slice')
       5.360.1ms       5.420.1ms     1.01  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       5.200.2ms       5.250.2ms     1.01  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       23.90.2ms       24.10.2ms     1.01  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'numpy_array_take_all_values')
       4.650.2ms       4.690.2ms     1.01  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       3.960.1ms      3.990.07ms     1.01  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
         61.81ms       62.20.5ms     1.01  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'numpy_array_take_all_values')
          3353ms          3363ms     1.00  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       6.910.1ms       6.930.2ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
          4808ms          4816ms     1.00  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       4.430.2ms       4.430.1ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       7.050.2ms       7.050.3ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       4.430.2ms       4.430.1ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       4.030.1ms       4.030.1ms     1.00  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
          3427ms          3424ms     1.00  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       3.550.1ms       3.550.2ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
         31.11ms         31.11ms     1.00  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'continuous_slice')
      3.020.06ms      3.010.05ms     1.00  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       47.60.3ms         47.41ms     1.00  benchmarks.TimeIndexing.time_loc([1000000, 10], 'slice')
       57.30.5ms         57.02ms     0.99  benchmarks.TimeIndexing.time_loc([1000000, 10], 'bool_array')
      3.560.06ms       3.530.2ms     0.99  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
          3445ms          3413ms     0.99  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       4.580.2ms       4.540.1ms     0.99  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       6.850.1ms      6.780.07ms     0.99  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       35.50.5ms       35.10.8ms     0.99  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       33.60.4ms       33.20.8ms     0.99  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
        7.390.3s        7.290.2s     0.99  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
      4.620.07ms       4.560.1ms     0.99  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
          4898ms          4825ms     0.99  benchmarks.TimeIndexingNumericSeries.time_getitem_lists([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
          3502ms          3443ms     0.98  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
       4.050.2ms       3.970.1ms     0.98  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       4.180.2ms       4.100.1ms     0.98  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
      5.420.07ms       5.300.2ms     0.98  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
      4.250.07ms       4.140.2ms     0.97  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       5.260.1ms       5.120.1ms     0.97  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
         55020ms         53320ms     0.97  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'slice')
       3.400.2ms      3.290.07ms     0.97  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       17.00.6ms       16.50.2ms     0.97  benchmarks.TimeIndexingColumns.time_loc([5000, 5000])
        7.300.2s        7.070.2s     0.97  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       4.650.1ms      4.500.04ms     0.97  benchmarks.TimeIndexing.time_loc([1000000, 10], 'scalar')
       3.200.2ms      3.090.08ms     0.97  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
         35.31ms         34.11ms     0.97  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       3.440.2ms      3.320.06ms     0.97  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
         84520ms         81660ms     0.97  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'function')
         84730ms         81730ms     0.97  benchmarks.TimeIndexing.time_loc([5000, 5000], 'function')
      3.080.08ms      2.970.07ms     0.96  benchmarks.TimeIndexingNumericSeries.time_iloc_list_like([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
      7.560.05ms       7.280.2ms     0.96  benchmarks.TimeIndexing.time_loc([1000000, 10], 'continuous_slice')
       3.650.1ms       3.520.1ms     0.96  benchmarks.TimeIndexingNumericSeries.time_loc_list_like([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       5.750.2ms       5.530.1ms     0.96  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
      3.970.07ms      3.820.08ms     0.96  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       17.70.3ms       17.00.1ms     0.96  benchmarks.TimeIndexing.time_iloc([5000, 5000], 'python_list_take_10_values')
      4.500.08ms       4.320.1ms     0.96  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       31.80.9ms       30.50.6ms     0.96  benchmarks.TimeIndexing.time_loc([5000, 5000], 'continuous_slice')
          3442ms          3294ms     0.96  benchmarks.TimeIndexingNumericSeries.time_getitem_array([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       6.940.3ms       6.650.2ms     0.96  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       7.910.4ms       7.580.3ms     0.96  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       4.960.2ms       4.740.3ms     0.96  benchmarks.TimeIndexingNumericSeries.time_iloc_scalar([100000, 1], <class 'numpy.uint64'>, 'unique_monotonic_inc')
      4.090.09ms       3.910.2ms     0.96  benchmarks.TimeIndexingNumericSeries.time_getitem_list_like([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       3.550.2ms      3.390.08ms     0.95  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       22.30.7ms       21.20.8ms     0.95  benchmarks.TimeIndexing.time_loc([5000, 5000], 'scalar')
       9.620.3ms       9.120.1ms     0.95  benchmarks.TimeIndexingNumericSeries.time_loc_array([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       3.790.1ms       3.580.2ms     0.95  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
       6.060.3ms       5.720.1ms     0.94  benchmarks.TimeIndexingNumericSeries.time_loc_scalar([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
         60120ms         56620ms     0.94  benchmarks.TimeIndexing.time_loc([5000, 5000], 'bool_array')
       7.290.2ms       6.870.1ms     0.94  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       5.860.4ms       5.520.1ms     0.94  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.float64'>, 'unique_monotonic_inc')
       6.980.1ms       6.570.2ms     0.94  benchmarks.TimeIndexingNumericSeries.time_iloc_array([100000, 1], <class 'numpy.int64'>, 'nonunique_monotonic_inc')
      3.520.04ms       3.310.1ms     0.94  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       7.520.2ms       7.020.2ms     0.93  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'continuous_slice')
       23.20.6ms       21.60.3ms     0.93  benchmarks.TimeIndexingColumns.time_loc([1000000, 10])
      3.620.08ms       3.380.1ms     0.93  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
      3.690.09ms       3.430.1ms     0.93  benchmarks.TimeIndexingNumericSeries.time_iloc_slice([100000, 1], <class 'numpy.uint64'>, 'nonunique_monotonic_inc')
       4.390.2ms       4.070.1ms     0.93  benchmarks.TimeIndexingNumericSeries.time_getitem_slice([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
       5.000.3ms       4.490.4ms    ~0.90  benchmarks.TimeIndexingNumericSeries.time_getitem_scalar([100000, 1], <class 'numpy.int64'>, 'unique_monotonic_inc')
       4.790.8ms       4.210.2ms    ~0.88  benchmarks.TimeIndexingNumericSeries.time_loc_slice([100000, 1], <class 'numpy.float64'>, 'nonunique_monotonic_inc')
-         3719ms          1833ms     0.49  benchmarks.TimeIndexing.time_iloc([1000000, 10], 'function')
-        37320ms          1814ms     0.49  benchmarks.TimeIndexing.time_loc([1000000, 10], 'function')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```
</details>